+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+# 0.2.0
+
+* Rename `tech_name` to `namespace`
+* Move `api` call to `@api true` module attribute, temporary make function as method name
+* Update to be compilebar on newest Elixir
+
 # 0.1.1
 
-* disable strict documentation requirement, add module attribute to enable it back
+* Disable strict documentation requirement, add module attribute to enable it back
 
 # 0.1.0